rename arquivos js de index p/ main
