+++ public/index.html
@@ -12,7 +12,7 @@
   <header>
     <nav>
       <ul>
-        <li><a href="/#">Coders</a></li>
+        <li><a href="/#">girlyHub</a></li>
         <li> <a href="/#geek">Geek</a></li>
         <li><a href="/#autocuidado">Autocuidado</a></li>
       </ul>
